Implementation of BH/VH
Implementation of ```BoundaryHandler()``` and ```VelocityHandler()``` allow the user to alter the behaviour of the optimisation algorithm. The example file has been updated to demonstrate how this behaviour can be changed with ease.

It is worth noting that this is not the most extensible method of implementing this functionality and this will be altered in a later commit to ensure that the user can develop their own methods to try new handler strategies.

Resolves: #3